third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/computed-properties-retain-expanded-expected.txt

Verifies that Computed Style preserves property expansion on re-rendering.

Inspected
Other

#inspected computed styles: 
color: rgb(0, 0, 255);
display: flex;
    flex - #inspected computed-proper…xpanded.html:5 -> computed-properties-retain-expanded.html:5:13
    OVERLOADED block - div computed-proper…panded.html:14 -> computed-properties-retain-expanded.html:14:6
    OVERLOADED block - div user agent stylesheet

#other computed styles: 
color: rgb(0, 0, 0);
display: inline;
    inline - #other computed-proper…panded.html:10 -> computed-properties-retain-expanded.html:10:9
    OVERLOADED block - div computed-proper…panded.html:14 -> computed-properties-retain-expanded.html:14:6
    OVERLOADED block - div user agent stylesheet
